feat: Add render with provider
